requirements.txt
# --- Core Web Framework ---
fastapi
uvicorn
gunicorn[uvloop]  # Added for production server

# --- AI & LangChain (Modern, compatible versions) ---
langchain
langchain-community
langchain-experimental
langchain-google-genai
langchain-neo4j
langchain-text-splitters
pypdf

# --- Databases ---
supabase
neo4j

# --- Utilities ---
python-dotenv
pydantic
tiktoken
python-dateutil
pytz  # Added for robust timezone handling
email_validator  # Added for email validation in Pydantic models
twilio
apscheduler

# --- Google Calendar API ---
google-api-python-client
google-auth-httplib2
google-auth-oauthlib